// src/app/_services/swagger-auto-generated/model/responseDataJwtTokenDto.ts
/**
 * API for controlling church music
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.1
 * Contact: kb.developer.br@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { JwtTokenDto } from './jwtTokenDto';
import { ResponseError } from './responseError';

export interface ResponseDataJwtTokenDto { 
    content?: JwtTokenDto;
    error?: ResponseError;
}